Refactor and integrate BagerTest and RouletteController
Removed duplicate BagerTest.cs and RouletteController.cs from the project root, consolidating their logic under Assets/[Game]/Scripts/Test/. Enhanced BagerTest and RouletteController with improved debug logging, more robust ball positioning, and tighter integration between the two scripts. Updated SampleScene.unity to reflect new serialized fields and references for improved roulette simulation and debugging.
